Fix app crash when the last recipe is deleted
Without `if (!recipe.name.isEmpty)` in RecipeRowView, the empty recipe is accessed even though the recipe has been deleted.
Moving the if to RecipesView still causes the issue.
After the last recipe is deleted, the fetched recipes do not become empty (though it should be). However with the change of this commit the app works fine.